+++ scripts/build.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+colcon build --packages-select calculator_pkg --symlink-install
+source install/setup.bash
