+++ scripts/entrypoint.sh
@@ -1,1 +1,3 @@
+#!/bin/bash
+
 echo "entrypoint.sh: Starting the application"
