+++ scripts/unit_test.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+
+# Run unit tests
+echo "Running unit tests"
+colcon test --packages-select calculator_pkg
+colcon test-result --verbose
+
+# Todo: add coverage report
